requirements.txt
torch
transformers
pillow
pandas
numpy
scikit-learn
mediapipe
opencv-python
xgboost